requirements.txt
Flask==2.1.1
pytest==7.4.0
werkzeug==2.0.3
requests==2.26.0
oauthlib==3.1.1
requests-oauthlib==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Dance[sqla]==3.3.1
